# apps/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "operationId": "EventsController_createEvent",
        "summary": "Ingest a new event",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEventResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate event (idempotency key conflict)"
          }
        },
        "tags": [
          "events"
        ]
      },
      "get": {
        "operationId": "EventsController_listEvents",
        "summary": "List events with filters and pagination",
        "parameters": [
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "inventory_change",
                "order_change",
                "calendar_event",
                "dashboard_update",
                "wine_update",
                "report_event",
                "notification_sent",
                "user_action",
                "system_event",
                "provider_change",
                "template_change"
              ],
              "type": "string"
            }
          },
          {
            "name": "sourcePage",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "dashboard",
                "inventory",
                "wine_library",
                "orders",
                "calendar",
                "reports",
                "communications",
                "providers",
                "documents",
                "notifications",
                "settings",
                "system"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns events list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/calendar/events": {
      "post": {
        "operationId": "CalendarController_createEvent",
        "summary": "Create a new calendar event",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calendar"
        ]
      },
      "get": {
        "operationId": "CalendarController_listEvents",
        "summary": "List calendar events with filters",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "delivery",
                "order",
                "meeting",
                "inventory",
                "tasting",
                "reminder",
                "recurring",
                "custom",
                "provider_birthday",
                "holiday",
                "delivery_eta",
                "provider_unavailable",
                "inventory_count",
                "high_volume_expected"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "approved",
                "dismissed",
                "completed",
                "cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeRecurring",
            "required": false,
            "in": "query",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns events list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/events/{eventId}": {
      "get": {
        "operationId": "CalendarController_getEvent",
        "summary": "Get a specific calendar event",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "calendar"
        ]
      },
      "patch": {
        "operationId": "CalendarController_updateEvent",
        "summary": "Update a calendar event",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "calendar"
        ]
      },
      "delete": {
        "operationId": "CalendarController_deleteEvent",
        "summary": "Delete a calendar event",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "this",
                "this_and_future",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/recurrence/{ruleId}/generate": {
      "post": {
        "operationId": "CalendarController_generateOccurrences",
        "summary": "Generate occurrences for a recurring event",
        "parameters": [
          {
            "name": "ruleId",
            "required": true,
            "in": "path",
            "description": "Recurrence rule ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "horizonDate",
            "required": false,
            "in": "query",
            "description": "Generate up to this date (YYYY-MM-DD)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Occurrences generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateOccurrencesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/recurrence/{ruleId}": {
      "get": {
        "operationId": "CalendarController_getRecurrenceRule",
        "summary": "Get a recurrence rule",
        "parameters": [
          {
            "name": "ruleId",
            "required": true,
            "in": "path",
            "description": "Recurrence rule ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the recurrence rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurrenceRuleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Rule not found"
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/event-types/{restaurantId}": {
      "get": {
        "operationId": "CalendarController_getEventTypes",
        "summary": "List event types for a restaurant",
        "parameters": [
          {
            "name": "restaurantId",
            "required": true,
            "in": "path",
            "description": "Restaurant UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event types list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/event-types": {
      "post": {
        "operationId": "CalendarController_createEventType",
        "summary": "Create a custom event type",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event type created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTypeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/event-types/{id}": {
      "patch": {
        "operationId": "CalendarController_updateEventType",
        "summary": "Update a custom event type",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event type ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event type updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTypeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calendar"
        ]
      },
      "delete": {
        "operationId": "CalendarController_deleteEventType",
        "summary": "Delete a custom event type",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event type ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event type deleted"
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/events/{eventId}/status": {
      "patch": {
        "operationId": "CalendarController_updateEventStatus",
        "summary": "Update event status",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "calendar"
        ]
      }
    },
...
  },
  "info": {
    "title": "WineOps AI API",
    "description": "\n## Overview\n\nReal-time API for Restaurant Wine Inventory & Procurement Management.\n\n### Features\n\n- **Dashboard Aggregation**: Single endpoint for all dashboard data (API Bus pattern)\n- **One-Tap Actions**: Quick approval workflows for managers\n- **Toast POS Integration**: Proxy endpoints for Toast API\n- **Real-time Updates**: WebSocket support for live data sync\n- **Rate Limiting**: Per-user and per-restaurant quotas\n\n### Authentication\n\nAll endpoints (except public ones) require a Bearer token in the Authorization header.\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n### Rate Limits\n\n| Endpoint Type | Limit | Window |\n|--------------|-------|--------|\n| Default | 100 | 60s |\n| Auth | 10 | 60s |\n| Upload | 10 | 300s |\n| AI | 20 | 60s |\n\n### WebSocket Events\n\nConnect to `ws://localhost:4000` and join rooms:\n- `restaurant:{restaurantId}` - Restaurant-wide events\n- `user:{userId}` - User-specific events\n\nEvents:\n- `notification` - New notifications\n- `inventory_update` - Inventory changes\n- `order_update` - Order status changes\n- `one_tap_action` - One-tap action updates\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "WineOps Team",
      "url": "https://wineops.ai",
      "email": "support@wineops.ai"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wineops.ai/license"
    }
  },
...
  "servers": [
    {
...
    },
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string",
            "enum": [
              "inventory_change",
              "order_change",
              "calendar_event",
              "dashboard_update",
              "wine_update",
              "report_event",
              "notification_sent",
              "user_action",
              "system_event",
              "provider_change",
              "template_change"
            ]
          },
          "sourcePage": {
            "type": "string",
            "enum": [
              "dashboard",
              "inventory",
              "wine_library",
              "orders",
              "calendar",
              "reports",
              "communications",
              "providers",
              "documents",
              "notifications",
              "settings",
              "system"
            ]
          },
          "payload": {
            "type": "object"
          },
          "schemaVersion": {
            "type": "number"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "eventType",
          "sourcePage",
          "payload"
        ]
      },
...
    }
  }
}